fixed warings

+++ IO/SocketUtils.h
@@ -8,7 +8,7 @@
 
 namespace IO
 {
-    void ReadWithProgress(int fd, int length, NBuffer &buffer, float *progress);
+    void ReadWithProgress(int fd, size_t length, NBuffer &buffer, float *progress);
 
     // Returns true if the fd became ready, false if it didn't or
     // if there was another error.
